// packages/react-select/src/types.d.ts
export interface OptionTypeBase {
  [key: string]: any;
}

export type OptionsType<OptionType extends OptionTypeBase> = ReadonlyArray<
  OptionType
>;

export interface GroupTypeBase<OptionType extends OptionTypeBase> {
  options: OptionsType<OptionType>;
  [key: string]: any;
}

export type GroupedOptionsType<
  OptionType extends OptionTypeBase,
  GroupType extends GroupTypeBase<OptionType> = GroupTypeBase<OptionType>
> = ReadonlyArray<GroupType>;

export type ValueType<
  OptionType extends OptionTypeBase,
  IsMulti extends boolean
> = IsMulti extends true ? OptionsType<OptionType> : OptionType | null;
// ...
export interface SelectOptionActionMeta<OptionType extends OptionTypeBase> {
  action: 'select-option';
  option: OptionType | undefined;
  name?: string;
}

export interface DeselectOptionActionMeta<OptionType extends OptionTypeBase> {
  action: 'deselect-option';
  option: OptionType | undefined;
  name?: string;
}

export interface RemoveValueActionMeta<OptionType extends OptionTypeBase> {
  action: 'remove-value';
  removedValue: OptionType;
  name?: string;
}

export interface PopValueActionMeta<OptionType extends OptionTypeBase> {
  action: 'pop-value';
  removedValue: OptionType;
  name?: string;
}

export interface ClearActionMeta {
  action: 'clear';
  name?: string;
}

export interface CreateOptionActionMeta {
  action: 'create-option';
  name?: string;
}

export type ActionMeta<OptionType extends OptionTypeBase> =
  | SelectOptionActionMeta<OptionType>
  | DeselectOptionActionMeta<OptionType>
  | RemoveValueActionMeta<OptionType>
  | PopValueActionMeta<OptionType>
  | ClearActionMeta
  | CreateOptionActionMeta;